json: Make json::validator publically usable
